+++ src/vs/workbench/services/textfile/test/browser/browserTextFileService.io.test.ts
@@ -47,10 +47,10 @@ if (isWeb) {
 
 				const collection = new ServiceCollection();
 				collection.set(IFileService, fileService);
-
-				collection.set(IWorkingCopyFileService, disposables.add(new WorkingCopyFileService(fileService, new WorkingCopyService(), instantiationService, new UriIdentityService(fileService))));
+				collection.set(IWorkingCopyFileService, disposables.add(new WorkingCopyFileService(fileService, disposables.add(new WorkingCopyService()), instantiationService, disposables.add(new UriIdentityService(fileService)))));
 
 				service = disposables.add(instantiationService.createChild(collection).createInstance(TestBrowserTextFileServiceWithEncodingOverrides));
+				disposables.add(<TextFileEditorModelManager>service.files);
 
 				await fileProvider.mkdir(URI.file(testDir));
 				for (const fileName in files) {
@@ -65,8 +65,6 @@ if (isWeb) {
 			},
 
 			teardown: async () => {
-				(<TextFileEditorModelManager>service.files).dispose();
-
 				disposables.clear();
 			},
