+++ src/vs/workbench/services/textfile/test/browser/textFileEditorModelManager.test.ts
@@ -18,18 +18,17 @@ import { DisposableStore } from 'vs/base/common/lifecycle';
 
 suite('Files - TextFileEditorModelManager', () => {
 
-	let disposables: DisposableStore;
+	const disposables = new DisposableStore();
 	let instantiationService: IInstantiationService;
 	let accessor: TestServiceAccessor;
 
 	setup(() => {
-		disposables = new DisposableStore();
 		instantiationService = workbenchInstantiationService(undefined, disposables);
 		accessor = instantiationService.createInstance(TestServiceAccessor);
 	});
 
 	teardown(() => {
-		disposables.dispose();
+		disposables.clear();
 	});
 
 	test('add, remove, clear, get, getAll', function () {
