+++ src/vs/workbench/services/textmodelResolver/test/browser/textModelResolverService.test.ts
@@ -22,21 +22,18 @@ import { DisposableStore } from 'vs/base/common/lifecycle';
 
 suite('Workbench - TextModelResolverService', () => {
 
-	let disposables: DisposableStore;
+	const disposables = new DisposableStore();
 	let instantiationService: IInstantiationService;
 	let accessor: TestServiceAccessor;
-	let model: TextFileEditorModel;
 
 	setup(() => {
-		disposables = new DisposableStore();
 		instantiationService = workbenchInstantiationService(undefined, disposables);
 		accessor = instantiationService.createInstance(TestServiceAccessor);
+		disposables.add(<TextFileEditorModelManager>accessor.textFileService.files);
 	});
 
 	teardown(() => {
-		model?.dispose();
-		(<TextFileEditorModelManager>accessor.textFileService.files).dispose();
-		disposables.dispose();
+		disposables.clear();
 	});
 
 	test('resolve resource', async () => {
