+++ src/vs/workbench/services/workingCopy/test/browser/fileWorkingCopyManager.test.ts
@@ -18,21 +18,20 @@ import { DisposableStore } from 'vs/base/common/lifecycle';
 
 suite('FileWorkingCopyManager', () => {
 
-	let disposables: DisposableStore;
+	const disposables = new DisposableStore();
 	let instantiationService: IInstantiationService;
 	let accessor: TestServiceAccessor;
 
 	let manager: IFileWorkingCopyManager<TestStoredFileWorkingCopyModel, TestUntitledFileWorkingCopyModel>;
 
 	setup(() => {
-		disposables = new DisposableStore();
 		instantiationService = workbenchInstantiationService(undefined, disposables);
 		accessor = instantiationService.createInstance(TestServiceAccessor);
 
 		accessor.fileService.registerProvider(Schemas.file, new TestInMemoryFileSystemProvider());
 		accessor.fileService.registerProvider(Schemas.vscodeRemote, new TestInMemoryFileSystemProvider());
 
-		manager = new FileWorkingCopyManager(
+		manager = disposables.add(new FileWorkingCopyManager(
 			'testFileWorkingCopyType',
 			new TestStoredFileWorkingCopyModelFactory(),
 			new TestUntitledFileWorkingCopyModelFactory(),
@@ -41,12 +40,11 @@ suite('FileWorkingCopyManager', () => {
 			accessor.filesConfigurationService, accessor.workingCopyService, accessor.notificationService,
 			accessor.workingCopyEditorService, accessor.editorService, accessor.elevatedFileService, accessor.pathService,
 			accessor.environmentService, accessor.dialogService, accessor.decorationsService
-		);
+		));
 	});
 
 	teardown(() => {
-		manager.dispose();
-		disposables.dispose();
+		disposables.clear();
 	});
 
 	test('onDidCreate, get, workingCopies', async () => {
