+++ src/vs/workbench/services/workingCopy/test/browser/workingCopyBackupTracker.test.ts
@@ -24,13 +24,11 @@ import { TestWorkingCopy } from 'vs/workbench/test/common/workbenchTestServices'
 import { CancellationToken } from 'vs/base/common/cancellation';
 import { timeout } from 'vs/base/common/async';
 import { BrowserWorkingCopyBackupTracker } from 'vs/workbench/services/workingCopy/browser/workingCopyBackupTracker';
-import { DisposableStore, dispose } from 'vs/base/common/lifecycle';
+import { DisposableStore } from 'vs/base/common/lifecycle';
 import { IWorkingCopyEditorHandler, IWorkingCopyEditorService } from 'vs/workbench/services/workingCopy/common/workingCopyEditorService';
 import { bufferToReadable, VSBuffer } from 'vs/base/common/buffer';
 import { isWindows } from 'vs/base/common/platform';
 import { Schemas } from 'vs/base/common/network';
-import { IWorkspaceTrustRequestService } from 'vs/platform/workspace/common/workspaceTrust';
-import { TestWorkspaceTrustRequestService } from 'vs/workbench/services/workspaces/test/common/testWorkspaceTrustService';
 
 suite('WorkingCopyBackupTracker (browser)', function () {
 	let accessor: TestServiceAccessor;
@@ -103,8 +101,6 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 
 		disposables.add(registerTestResourceEditor());
 
-		instantiationService.stub(IWorkspaceTrustRequestService, new TestWorkspaceTrustRequestService(false));
-
 		const editorService: EditorService = disposables.add(instantiationService.createInstance(EditorService));
 		instantiationService.stub(IEditorService, editorService);
 
@@ -166,7 +162,7 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 		}
 
 		const resource = toResource.call(this, '/path/custom.txt');
-		const customWorkingCopy = new TestBackupWorkingCopy(resource);
+		const customWorkingCopy = disposables.add(new TestBackupWorkingCopy(resource));
 
 		// Normal
 		customWorkingCopy.setDirty(true);
@@ -192,17 +188,13 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 		assert.strictEqual(tracker.pendingBackupOperationCount, 1);
 		await workingCopyBackupService.joinDiscardBackup();
 		assert.strictEqual(workingCopyBackupService.hasBackupSync(customWorkingCopy), false);
-
-		customWorkingCopy.dispose();
 	});
 
-	async function restoreBackupsInit(): Promise<[TestWorkingCopyBackupTracker, TestServiceAccessor, DisposableStore]> {
+	async function restoreBackupsInit(): Promise<[TestWorkingCopyBackupTracker, TestServiceAccessor]> {
 		const fooFile = URI.file(isWindows ? 'c:\\Foo' : '/Foo');
 		const barFile = URI.file(isWindows ? 'c:\\Bar' : '/Bar');
 		const untitledFile1 = URI.from({ scheme: Schemas.untitled, path: 'Untitled-1' });
 		const untitledFile2 = URI.from({ scheme: Schemas.untitled, path: 'Untitled-2' });
-
-		const disposables = new DisposableStore();
 
 		const workingCopyBackupService = disposables.add(new InMemoryTestWorkingCopyBackupService());
 		const instantiationService = workbenchInstantiationService(undefined, disposables);
@@ -210,8 +202,6 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 
 		const part = await createEditorPart(instantiationService, disposables);
 		instantiationService.stub(IEditorGroupsService, part);
-
-		instantiationService.stub(IWorkspaceTrustRequestService, new TestWorkspaceTrustRequestService(false));
 
 		const editorService: EditorService = disposables.add(instantiationService.createInstance(EditorService));
 		instantiationService.stub(IEditorService, editorService);
@@ -233,11 +223,11 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 
 		accessor.lifecycleService.phase = LifecyclePhase.Restored;
 
-		return [tracker, accessor, disposables];
+		return [tracker, accessor];
 	}
 
 	test('Restore backups (basics, some handled)', async function () {
-		const [tracker, accessor, disposables] = await restoreBackupsInit();
+		const [tracker, accessor] = await restoreBackupsInit();
 
 		assert.strictEqual(tracker.getUnrestoredBackups().size, 0);
 
@@ -275,12 +265,10 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 			assert.ok(editor instanceof TestUntitledTextEditorInput);
 			assert.strictEqual(editor.resolved, true);
 		}
-
-		dispose(disposables);
 	});
 
 	test('Restore backups (basics, none handled)', async function () {
-		const [tracker, accessor, disposables] = await restoreBackupsInit();
+		const [tracker, accessor] = await restoreBackupsInit();
 
 		await tracker.testRestoreBackups({
 			handles: workingCopy => false,
@@ -290,12 +278,10 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 
 		assert.strictEqual(accessor.editorService.count, 0);
 		assert.strictEqual(tracker.getUnrestoredBackups().size, 4);
-
-		dispose(disposables);
 	});
 
 	test('Restore backups (basics, error case)', async function () {
-		const [tracker, , disposables] = await restoreBackupsInit();
+		const [tracker] = await restoreBackupsInit();
 
 		try {
 			await tracker.testRestoreBackups({
@@ -308,12 +294,10 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 		}
 
 		assert.strictEqual(tracker.getUnrestoredBackups().size, 4);
-
-		dispose(disposables);
 	});
 
 	test('Restore backups (multiple handlers)', async function () {
-		const [tracker, accessor, disposables] = await restoreBackupsInit();
+		const [tracker, accessor] = await restoreBackupsInit();
 
 		const firstHandler = tracker.testRestoreBackups({
 			handles: workingCopy => {
@@ -349,12 +333,10 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 			assert.ok(editor instanceof TestUntitledTextEditorInput);
 			assert.strictEqual(editor.resolved, true);
 		}
-
-		dispose(disposables);
 	});
 
 	test('Restore backups (editors already opened)', async function () {
-		const [tracker, accessor, disposables] = await restoreBackupsInit();
+		const [tracker, accessor] = await restoreBackupsInit();
 
 		assert.strictEqual(tracker.getUnrestoredBackups().size, 0);
 
@@ -399,7 +381,5 @@ suite('WorkingCopyBackupTracker (browser)', function () {
 				assert.strictEqual(editor.resolved, true);
 			}
 		}
-
-		dispose(disposables);
 	});
 });
