+++ src/vs/workbench/services/workingCopy/test/electron-sandbox/workingCopyHistoryTracker.test.ts
@@ -5,14 +5,14 @@
 
 import * as assert from 'assert';
 import { Event } from 'vs/base/common/event';
-import { TestContextService, TestStorageService, TestWorkingCopy } from 'vs/workbench/test/common/workbenchTestServices';
+import { TestContextService, TestWorkingCopy } from 'vs/workbench/test/common/workbenchTestServices';
 import { randomPath } from 'vs/base/common/extpath';
 import { join } from 'vs/base/common/path';
 import { URI } from 'vs/base/common/uri';
 import { WorkingCopyHistoryTracker } from 'vs/workbench/services/workingCopy/common/workingCopyHistoryTracker';
 import { WorkingCopyService } from 'vs/workbench/services/workingCopy/common/workingCopyService';
 import { UriIdentityService } from 'vs/platform/uriIdentity/common/uriIdentityService';
-import { TestEnvironmentService, TestFileService, TestLifecycleService, TestPathService, TestRemoteAgentService } from 'vs/workbench/test/browser/workbenchTestServices';
+import { TestFileService, TestPathService } from 'vs/workbench/test/browser/workbenchTestServices';
 import { DeferredPromise } from 'vs/base/common/async';
 import { IFileService } from 'vs/platform/files/common/files';
 import { Schemas } from 'vs/base/common/network';
@@ -25,44 +25,9 @@ import { CancellationToken } from 'vs/base/common/cancellation';
 import { IWorkingCopyHistoryEntry, IWorkingCopyHistoryEntryDescriptor } from 'vs/workbench/services/workingCopy/common/workingCopyHistory';
 import { assertIsDefined } from 'vs/base/common/types';
 import { VSBuffer } from 'vs/base/common/buffer';
-import { InMemoryFileSystemProvider } from 'vs/platform/files/common/inMemoryFilesystemProvider';
 import { DisposableStore } from 'vs/base/common/lifecycle';
-import { NativeWorkingCopyHistoryService } from 'vs/workbench/services/workingCopy/common/workingCopyHistoryService';
-import { NullLogService } from 'vs/platform/log/common/log';
-import { FileService } from 'vs/platform/files/common/fileService';
-import { LabelService } from 'vs/workbench/services/label/common/labelService';
 import { ensureNoDisposablesAreLeakedInTestSuite } from 'vs/base/test/common/utils';
-
-class TestWorkingCopyHistoryService extends NativeWorkingCopyHistoryService {
-
-	readonly _fileService: IFileService;
-	readonly _configurationService: TestConfigurationService;
-	readonly _lifecycleService: TestLifecycleService;
-
-	constructor() {
-		const environmentService = TestEnvironmentService;
-		const logService = new NullLogService();
-		const fileService = new FileService(logService);
-
-		fileService.registerProvider(Schemas.vscodeUserData, new InMemoryFileSystemProvider());
-
-		const remoteAgentService = new TestRemoteAgentService();
-
-		const uriIdentityService = new UriIdentityService(fileService);
-
-		const labelService = new LabelService(environmentService, new TestContextService(), new TestPathService(), new TestRemoteAgentService(), new TestStorageService(), new TestLifecycleService());
-
-		const lifecycleService = new TestLifecycleService();
-
-		const configurationService = new TestConfigurationService();
-
-		super(fileService, remoteAgentService, environmentService, uriIdentityService, labelService, lifecycleService, logService, configurationService);
-
-		this._fileService = fileService;
-		this._configurationService = configurationService;
-		this._lifecycleService = lifecycleService;
-	}
-}
+import { TestWorkingCopyHistoryService } from 'vs/workbench/services/workingCopy/test/electron-sandbox/workingCopyHistoryService.test';
 
 suite('WorkingCopyHistoryTracker', () => {
 
@@ -106,12 +71,10 @@ suite('WorkingCopyHistoryTracker', () => {
 		historyHome = joinPath(testDir, 'User', 'History');
 		workHome = joinPath(testDir, 'work');
 
-		workingCopyHistoryService = disposables.add(new TestWorkingCopyHistoryService());
+		workingCopyHistoryService = disposables.add(new TestWorkingCopyHistoryService(disposables));
 		workingCopyService = disposables.add(new WorkingCopyService());
 		fileService = workingCopyHistoryService._fileService;
 		configurationService = workingCopyHistoryService._configurationService;
-
-		disposables.add(fileService.registerProvider(Schemas.inMemory, new InMemoryFileSystemProvider()));
 
 		tracker = disposables.add(createTracker());
