+++ src/vs/workbench/test/browser/parts/editor/editorDiffModel.test.ts
@@ -15,18 +15,17 @@ import { DisposableStore } from 'vs/base/common/lifecycle';
 
 suite('TextDiffEditorModel', () => {
 
-	let disposables: DisposableStore;
+	const disposables = new DisposableStore();
 	let instantiationService: IInstantiationService;
 	let accessor: TestServiceAccessor;
 
 	setup(() => {
-		disposables = new DisposableStore();
 		instantiationService = workbenchInstantiationService(undefined, disposables);
 		accessor = instantiationService.createInstance(TestServiceAccessor);
 	});
 
 	teardown(() => {
-		disposables.dispose();
+		disposables.clear();
 	});
 
 	test('basics', async () => {
